+++ examples/blog/main.py
@@ -11,12 +11,7 @@ from auth import (
 )
 from fastapi import Depends, FastAPI, HTTPException, Query, status
 from fastapi.security import OAuth2PasswordRequestForm
-from models import (
-    MongoPost,
-    RedisPost,
-    SqlInternalAuthor,
-    SqlPost,
-)
+from models import MongoPost, RedisPost, SqlInternalAuthor, SqlPost
 from pydantic import BaseModel
 from schemas import InternalAuthor, Post, TokenResponse
 from stores import MongoStoreDep, RedisStoreDep, SqlStoreDep, clear_stores