+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Allowed `EmbeddedJsonModel`'s to also contain `EmbeddedJsonModel` if there is need.
 - Allowed extra key-word args to be passed to the `SQLModel()`
+- Made the defining of fields by calling `Field()` or `Relationship()` mandatory.
+  This is because SQLModel's require this, and fail if this is not the case.
+- Add support for "Many-to-One" relationships in SQL `insert()` implementation
 
 ### Added